issue 165 changed to place all defines in one file
