--- examples/dependencies/concatexample/concat_a.js
@@ -1,10 +1,0 @@
-var concat_a = function() {
-  var module = "concat_a";
-  
-  return {
-    getModuleName: function() {
-      return module;
-    }
-  }
-}
-exports.concat_a = concat_a;
