--- examples/dependencies/concatexample/concat_b.js
@@ -1,10 +1,0 @@
-var concat_b = function() {
-  var module = "concat_b";
-
-  return {
-    getModuleName: function() {
-      return module;
-    }
-  }
-}
-exports.concat_b = concat_b;
