--- examples/dependencies/concatexample/concat_c.js
@@ -1,9 +1,0 @@
-var concat_c = function() {
-	var module = "concat_c";
-	return {
-		getModuleName: function() {
-			return module;
-		}
-	}
-}
-exports.concat_c = concat_c;
